feat(captive): Adds user input capture and storage
Implements the ability to capture user inputs from the captive portal's HTML form and store them.
Displays the captured user credentials via a notification.
Adds functionality to store the captured data to a file on the SD card if present.
Creates directories for portals and user data on the SD card upon formatting.