feat(captive): add preference to dump config to SD
Allows users to select whether or not to dump the captive portal configuration to the SD card.

This adds a new "Preferences" option to the main menu, where users can choose between "Dump to SD" and "No dump". The selected preference is stored and used to determine whether to create and append the configuration data to the SD card.