Refactor: BLE scanner for improved reliability
Removes the timer task and relies on the scan timeout event to restart scanning, leading to a more reliable and simplified BLE scanning process.
The scan duration is increased to 10 minutes.
Also improves the UART sender to correctly send BLE packets.
The adv scanner module is refactored to use the general submenu component.