feat: Adds captive portal redirect functionality
Implements a redirect feature for the captive portal, allowing users to be redirected to a specific page after submitting their credentials.
This includes adding a new "redirect" option to the main menu, allowing the user to specify a redirect HTML file from the SD card. A default redirect page is also included. After a 5 second delay, the user is redirected to the root page.